pyproject.toml
[project]
name = "opentulpa"
version = "0.1.0"
description = "A background-capable AI agent using LangGraph with mem0 memory"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "kvyb", email = "you@example.com" }
]
license = { text = "MIT" }
dependencies = [
    "langchain>=0.3",
    "langgraph>=0.2",
    "langchain-openai>=0.3",
    "langgraph-checkpoint-sqlite>=2.0",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
    "crawl4ai>=0.7.0",
    "mem0ai[llms]>=1.0",
    "pypdf>=5.0",
    "apscheduler>=3.10",
    "slack-sdk>=3.27",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opentulpa"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
